Implement signed_size_t type
This type was implemented as follows:
- ptrdiff_t was replaced with signed_size_t via sed(1).
  - signed_size_t started as a 32-bit int typedef.
- The code was compiled with Clang 21.1.8 using:
    -Wall -Wextra -Wsign-compare -Wshorten-64-to-32 -Wsign-conversion -Wimplicit-int-conversion
- All instances of 64-to-32 bit shortening occuring at pointer
  arithmetic sites (caught by clang) were fixed by adding
  ptrdiff_t back in judiciously.
- signed_size_t was redefined to a 64-bit integer type and
  ptrdiff_t was redefined to a 32-bit type.
- After another compile, subsequent warnings were fixed appropriately.
- Additional touchup fixes were applied + some minor cleanup.
- Regression tests were run yet again.

While this did result in some degree of separation between signed sizes
and pointer differences/offsets, ptrdiff_t still had to be used quite
liberally (wc(1) counts a remaining 567 instances of the type).
Not all uses of ptrdiff_t are present solely for variables holding just
pointer differences; often variables will be used for other things too.
(AT&T code frequently repurposes variables like 'i' and 'n' in ways that
are difficult to refactor.) Other times, a variable that ultimately
holds a pointer offset will not immediately be obvious at a glance due
to code complexity (hence why Clang is absolutely necessary for static
analysis). In other cases sizes will be calculated using pointer
offsets, and while ptrdiff_t and size_t are _usually the same size,
some memory models will have a ptrdiff_t that is larger than the size_t,
so for correctness that has been taken into minor consideration (i.e.,
the possibility of a signed_size_t having a lower bitness means some
sections of code use ptrdiff_t when calculating sizes to avoid
potentially truncating the result).

Part of https://github.com/ksh93/ksh/pull/867

+++ src/lib/libast/sfio/sfmove.c
@@ -150,7 +150,7 @@ Sfoff_t sfmove(Sfio_t*	fr,	/* moving data from this stream */
 						r = (ssize_t)n;
 				}
 				else	r = -1;
-				if((r = (ssize_t)SFFILBUF(fr,r)) <= 0)
+				if((r = (ssize_t)SFFILBUF(fr,(ptrdiff_t)r)) <= 0)
 					break;
 				next = fr->next;
 			}
